+++ CHANGES.md
@@ -5,6 +5,7 @@ Markdown Writer FX Change Log
 
 - Project specific options: Select "Store in project" checkbox in Options dialog
   to save options in per project (in file `<project-root>/.markdownwriterfx`).
+- "Format only modified paragraphs on Save" checkbox added to Options dialog.
 - Disabled "Save As" for read-only editors (binary or too-large files).